Małe zmiany
- Zakres wartości na silniki od 0 do 1000 (a nie od 0 do 2000)
- Nowa funkcja ustawiania prędkości silników

+++ FlightController_v1_0/Sensors.h
@@ -28,8 +28,8 @@ class SensorsClass
 	
 // === ZMIENNE ===
  public:
-	float pitch = 0;
-	float roll = 0;
+	double pitch = 0;
+	double roll = 0;
 	float yaw = 0;
 	double dt_;
